requirements.txt
tensorflow
protobuf
Pillow
numpy
